extract class and functions from boswatch.py to includes
